refactor: meta-service: simplify raft store and state machine
- Remove `is_opened` flag from `RaftStore`

- Remove obsolete config no_sync, which is only used by sled tree store

- Add config to MetaRaftLog

- Remove `RaftStoreInner`
