+++ src/meta/service/src/api/grpc/grpc_service.rs
@@ -413,7 +413,7 @@ impl MetaService for MetaServiceImpl {
 
         let meta_node = self.try_get_meta_node()?;
 
-        let strm = meta_node.raft_store.inner().export();
+        let strm = meta_node.raft_store.clone().export();
 
         let chunk_size = 32;
         // - Chunk up upto 32 Ok items inside a Vec<String>;
@@ -446,7 +446,7 @@ impl MetaService for MetaServiceImpl {
 
         let meta_node = self.try_get_meta_node()?;
 
-        let strm = meta_node.raft_store.inner().export();
+        let strm = meta_node.raft_store.clone().export();
 
         let chunk_size = request.get_ref().chunk_size.unwrap_or(32) as usize;
         // - Chunk up upto `chunk_size` Ok items inside a Vec<String>;
