+++ src/meta/service/tests/it/meta_node/meta_node_lifecycle.rs
@@ -832,7 +832,7 @@ async fn test_meta_node_restart_single_node() -> anyhow::Result<()> {
             .await?;
         log_index += 1;
 
-        want_hs = leader.raft_store.log.clone().read_vote().await?;
+        want_hs = leader.raft_store.log().clone().read_vote().await?;
 
         leader.stop().await?;
     }
@@ -866,7 +866,7 @@ async fn test_meta_node_restart_single_node() -> anyhow::Result<()> {
 
     info!("--- check hard state");
     {
-        let hs = leader.raft_store.log.clone().read_vote().await?;
+        let hs = leader.raft_store.log().clone().read_vote().await?;
         assert_eq!(want_hs, hs);
     }
 
@@ -874,7 +874,7 @@ async fn test_meta_node_restart_single_node() -> anyhow::Result<()> {
     {
         let logs = leader
             .raft_store
-            .log
+            .log()
             .clone()
             .try_get_log_entries(..)
             .await?;
