+++ src/meta/service/tests/it/store.rs
@@ -55,7 +55,7 @@ impl StoreBuilder<TypeConfig, LogStore, SMStore, MetaSrvTestContext> for MetaSto
         let sto = RaftStore::open(&tc.config.raft_config)
             .await
             .expect("fail to create store");
-        Ok((tc, sto.log.clone(), sto.state_machine.clone()))
+        Ok((tc, sto.log().clone(), sto.state_machine().clone()))
     }
 }
 
@@ -81,9 +81,9 @@ async fn test_meta_store_purge_cache() -> anyhow::Result<()> {
     {
         let sto = RaftStore::open(&tc.config.raft_config).await?;
 
-        sto.log.clone().save_vote(&Vote::new(10, 5)).await?;
+        sto.log().clone().save_vote(&Vote::new(10, 5)).await?;
 
-        sto.log
+        sto.log()
             .clone()
             .blocking_append([
                 Entry::new_blank(log_id(1, 2, 1)),
@@ -94,11 +94,11 @@ async fn test_meta_store_purge_cache() -> anyhow::Result<()> {
             ])
             .await?;
 
-        let stat = sto.log.read().await.stat();
+        let stat = sto.log().read().await.stat();
         assert_eq!(stat.payload_cache_item_count, 5);
 
         {
-            let r = sto.log.read().await;
+            let r = sto.log().read().await;
             let got = r.dump().write_to_string()?;
             println!("dump: {}", got);
             let want_dumped = r#"RaftLog:
@@ -120,13 +120,13 @@ ChunkId(00_000_000_000_000_000_200)
         // When purging up to index=4, all entries in the last open chunk will still be cached.
         // All previous entries are purge, although the cache is not full.
 
-        sto.log.clone().purge(log_id(1, 2, 4)).await?;
+        sto.log().clone().purge(log_id(1, 2, 4)).await?;
 
-        let r = sto.log.read().await;
+        let r = sto.log().read().await;
         let got = r.dump().write_to_string()?;
         println!("dump: {}", got);
 
-        let stat = sto.log.read().await.stat();
+        let stat = sto.log().read().await.stat();
         println!("stat: {:#}", stat);
         assert_eq!(stat.payload_cache_item_count, 3);
     }
@@ -149,24 +149,23 @@ async fn test_meta_store_restart() -> anyhow::Result<()> {
     {
         let sto = RaftStore::open(&tc.config.raft_config).await?;
         assert_eq!(id, sto.id);
-        assert!(!sto.is_opened);
-        assert_eq!(None, sto.log.clone().read_vote().await?);
+        assert_eq!(None, sto.log().clone().read_vote().await?);
 
         info!("--- update metasrv");
 
-        sto.log.clone().save_vote(&Vote::new(10, 5)).await?;
+        sto.log().clone().save_vote(&Vote::new(10, 5)).await?;
 
-        sto.log
+        sto.log()
             .clone()
             .blocking_append([Entry::new_blank(log_id(1, 2, 1))])
             .await?;
 
-        sto.log
+        sto.log()
             .clone()
             .save_committed(Some(log_id(1, 2, 2)))
             .await?;
 
-        sto.state_machine
+        sto.state_machine()
             .clone()
             .apply([Entry::new_blank(log_id(1, 2, 2))])
             .await?;
@@ -176,17 +175,16 @@ async fn test_meta_store_restart() -> anyhow::Result<()> {
     {
         let sto = RaftStore::open(&tc.config.raft_config).await?;
         assert_eq!(id, sto.id);
-        assert!(sto.is_opened);
-        assert_eq!(Some(Vote::new(10, 5)), sto.log.clone().read_vote().await?);
+        assert_eq!(Some(Vote::new(10, 5)), sto.log().clone().read_vote().await?);
 
-        assert_eq!(log_id(1, 2, 1), sto.log.clone().get_log_id(1).await?);
+        assert_eq!(log_id(1, 2, 1), sto.log().clone().get_log_id(1).await?);
         assert_eq!(
             Some(log_id(1, 2, 2)),
-            sto.log.clone().read_committed().await?
+            sto.log().clone().read_committed().await?
         );
         assert_eq!(
             None,
-            sto.state_machine.clone().applied_state().await?.0,
+            sto.state_machine().clone().applied_state().await?.0,
             "state machine is not persisted"
         );
     }
@@ -209,10 +207,10 @@ async fn test_meta_store_build_snapshot() -> anyhow::Result<()> {
 
     let (logs, want) = snapshot_logs();
 
-    sto.log.clone().blocking_append(logs.clone()).await?;
+    sto.log().clone().blocking_append(logs.clone()).await?;
     sto.state_machine().get_inner().apply_entries(logs).await?;
 
-    let curr_snap = sto.state_machine.clone().build_snapshot().await?;
+    let curr_snap = sto.state_machine().clone().build_snapshot().await?;
     assert_eq!(Some(new_log_id(1, 0, 9)), curr_snap.meta.last_log_id);
 
     info!("--- check snapshot");
@@ -227,12 +225,12 @@ async fn test_meta_store_build_snapshot() -> anyhow::Result<()> {
 
     info!("--- rebuild other 4 times, keeps only last 3");
     {
-        sto.state_machine.clone().build_snapshot().await?;
-        sto.state_machine.clone().build_snapshot().await?;
-        sto.state_machine.clone().build_snapshot().await?;
-        sto.state_machine.clone().build_snapshot().await?;
+        sto.state_machine().clone().build_snapshot().await?;
+        sto.state_machine().clone().build_snapshot().await?;
+        sto.state_machine().clone().build_snapshot().await?;
+        sto.state_machine().clone().build_snapshot().await?;
 
-        let snapshot_store = sto.state_machine.snapshot_store();
+        let snapshot_store = sto.state_machine().snapshot_store();
         let loader = snapshot_store.new_loader();
         let (snapshot_ids, _) = loader.load_snapshot_ids().await?;
         assert_eq!(3, snapshot_ids.len());
@@ -257,18 +255,18 @@ async fn test_meta_store_current_snapshot() -> anyhow::Result<()> {
 
     let (logs, want) = snapshot_logs();
 
-    sto.log.clone().blocking_append(logs.clone()).await?;
+    sto.log().clone().blocking_append(logs.clone()).await?;
     {
         let sm = sto.state_machine();
         sm.get_inner().apply_entries(logs).await?;
     }
 
-    sto.state_machine.clone().build_snapshot().await?;
+    sto.state_machine().clone().build_snapshot().await?;
 
     info!("--- check get_current_snapshot");
 
     let curr_snap = sto
-        .state_machine
+        .state_machine()
         .clone()
         .get_current_snapshot()
         .await?
@@ -307,10 +305,10 @@ async fn test_meta_store_install_snapshot() -> anyhow::Result<()> {
 
         info!("--- feed logs and state machine");
 
-        sto.log.clone().blocking_append(logs.clone()).await?;
+        sto.log().clone().blocking_append(logs.clone()).await?;
         sto.state_machine().get_inner().apply_entries(logs).await?;
 
-        snap = sto.state_machine.clone().build_snapshot().await?;
+        snap = sto.state_machine().clone().build_snapshot().await?;
     }
 
     let data = snap.snapshot;
@@ -323,7 +321,7 @@ async fn test_meta_store_install_snapshot() -> anyhow::Result<()> {
 
         info!("--- install snapshot");
         {
-            sto.state_machine
+            sto.state_machine()
                 .clone()
                 .do_install_snapshot(data.clone())
                 .await?;
@@ -356,7 +354,7 @@ async fn test_meta_store_install_snapshot() -> anyhow::Result<()> {
 
         info!("--- check snapshot");
         {
-            let curr_snap = sto.state_machine.clone().build_snapshot().await?;
+            let curr_snap = sto.state_machine().clone().build_snapshot().await?;
             let data = curr_snap.snapshot;
             let res = db_to_lines(&data).await?;
