+++ src/meta/store/src/local.rs
@@ -113,12 +113,6 @@ impl LocalMetaService {
         let raft_port = next_port();
         let mut config = configs::Config::default();
 
-        // On mac File::sync_all() takes 10 ms ~ 30 ms, 500 ms at worst, which very likely to fail a test.
-        if cfg!(target_os = "macos") {
-            warn!("Disabled fsync for meta service. fsync on mac is quite slow");
-            config.raft_config.no_sync = true;
-        }
-
         config.raft_config.id = 0;
 
         config.raft_config.config_id = raft_port.to_string();
